Markov Chain Feature: Part 2
This is pretty much another big rewrite. All of this because interfaces.

I'm an indecisive ass, so 'modes' from the previous commit are now called 'generators'.
The classes previously classified as 'generators' where used for creating buzzword sentences; consequently, those are now called 'buzzwords'.

In short, almost everything has been revamped to support the new, more generic generator system that can be accessed from anywhere using DI.
Currently, the two generators are BuzzwordGenerator and MarkovGenerator, the last of which contains all the new non-generator system features.
There are a few bugs with this system and I'm not happy with that. I'll probably re-rewrite this whole thing again later.